requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
cffi==1.16.0
cryptography==42.0.2
Django==4.2.9
django-cors-headers==4.3.1
django-rest-knox==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pillow==10.2.0
pycparser==2.21
PyJWT==2.8.0
python-environ==0.4.54
pytz==2023.4
sqlparse==0.4.4
typing-extensions==4.9.0
